refactor: relocate Apple plugin, interactor, interactions under platforms/apple (#975)
Phase 3 d.1. Makes getInteractor's core -> platforms routing the final shape by
moving the Apple-specific plugin pieces under src/platforms/apple/ while keeping
the generic registry where non-interactor core code can still import it.

Moved to src/platforms/apple/:
- plugin.ts       - the applePlugin instance (APPLE_SUPPORTS_BY_DEFAULT closures,
                    appLog/perf facets, createInteractor/discoverDevices), extracted
                    from the former core/platform-plugin/register-builtins.ts
- interactor.ts   - was core/interactors/apple.ts (createAppleInteractor)
- interactions.ts - was platforms/ios/interactions.ts (the Apple interaction
                    dispatcher: iOS synthesized gesture / tvOS remote-press /
                    macOS desktop-scroll); platforms/ios/ is now removed
- __tests__/watchos-sentinel.test.ts - co-located with its subject

Stayed in core/ (layering):
- The generic registry + PlatformPlugin type stay in core/platform-plugin/plugin.ts.
  core/capabilities.ts (non-interactor core) imports getPlugin/tryGetPlugin, and R3
  forbids core outside core/interactors/ from statically importing platforms/, so the
  registry cannot move.
- register-builtins.ts moved to core/interactors/register-builtins.ts (still core/):
  the android/linux/web wiring plus the registry-population entry point. As an
  interactor-seam module it is the one place R3 permits a static value import of
  platforms/, so it pulls in applePlugin and keeps the exhaustiveness assertion.
- apple-os-capabilities.ts stays in core/ (smallest move; the moved Apple closures
  reach it via a legal platforms -> core import).

Layering guard passes (R1/R2/R3, 678 files). The applePlugin only reaches leaf code
via lazy dynamic import(), so the new static platforms/apple/plugin.ts import at the
seam does not regress CLI cold-start.
