+++ src/daemon/__tests__/perf-plugin-routing-parity.test.ts
@@ -19,7 +19,7 @@ import {
   WEB_DESKTOP_DEVICE,
 } from '../../__tests__/test-utils/index.ts';
 import { getPlugin } from '../../core/platform-plugin/plugin.ts';
-import { registerBuiltinPlatformPlugins } from '../../core/platform-plugin/register-builtins.ts';
+import { registerBuiltinPlatformPlugins } from '../../core/interactors/register-builtins.ts';
 import { buildPerfResponseData } from '../handlers/session-perf.ts';
 import { PERF_UNAVAILABLE_REASON } from '../handlers/session-startup-metrics.ts';
