+++ src/platforms/apple/plugin.ts
@@ -1,11 +1,11 @@
-import { appleOsCapabilities } from './apple-os-capabilities.ts';
-import { registerPlatformPlugin, type PlatformPlugin } from './plugin.ts';
+import { appleOsCapabilities } from '../../core/platform-plugin/apple-os-capabilities.ts';
+import type { PlatformPlugin } from '../../core/platform-plugin/plugin.ts';
 import { PUBLIC_COMMANDS } from '../../command-catalog.ts';
 import { isAudioProbeSupportedDevice } from '../../kernel/audio-probe-support.ts';
-import { shouldUseHostMacFastPath, WEB_DESKTOP_DEVICE } from '../platform-inventory.ts';
-import type { Platform, DeviceInfo } from '../../kernel/device.ts';
-import type { DeviceInventoryRequest } from '../platform-inventory.ts';
-import type { RunnerContext } from '../interactor-types.ts';
+import { shouldUseHostMacFastPath } from '../../core/platform-inventory.ts';
+import type { DeviceInfo } from '../../kernel/device.ts';
+import type { DeviceInventoryRequest } from '../../core/platform-inventory.ts';
+import type { RunnerContext } from '../../core/interactor-types.ts';
 
 // ---------------------------------------------------------------------------
 // Apple family per-command capability closures. Originally RELOCATED VERBATIM from
@@ -108,15 +108,15 @@ const APPLE_UNSUPPORTED_HINT_BY_DEFAULT: Record<
   'transform-gesture': synthesisGestureUnsupportedHint,
 };
 
-// Each plugin WRAPS today's existing factories (src/core/interactors/*) and the
-// inventory if-chain (src/core/platform-inventory.ts) as LAZY methods. No leaf
-// code is rewritten: the dynamic `import()`s and the per-platform list calls are
-// byte-for-byte the same as the hand-authored `getInteractor` switch arms and
-// `listLocalDeviceInventory` branches. `as const satisfies PlatformPlugin`
-// preserves each plugin's literal `platforms` tuple so the totality assertion
-// below is a real compile-time check.
+// The Apple plugin WRAPS today's existing factories (its `createInteractor` in
+// `./interactor.ts`) and the inventory if-chain (src/core/platform-inventory.ts) as
+// LAZY methods. No leaf code is rewritten: the dynamic `import()`s and the per-platform
+// list calls are byte-for-byte the same as the hand-authored `getInteractor` switch arm
+// and `listLocalDeviceInventory` branch. `as const satisfies PlatformPlugin` preserves
+// the plugin's literal `platforms` tuple so the registry totality assertion (in
+// `core/interactors/register-builtins.ts`) is a real compile-time check.
 
-const applePlugin = {
+export const applePlugin = {
   id: 'apple',
   // Apple owns BOTH leaf platforms today — mirrors `case 'ios': case 'macos':`.
   platforms: ['ios', 'macos'],
@@ -140,116 +140,20 @@ const applePlugin = {
   // (ios/macos, any kind/target) reports perf-metrics support.
   perf: { supportsMetrics: () => true },
   createInteractor: async (device: DeviceInfo, runner: RunnerContext) => {
-    const { createAppleInteractor } = await import('../interactors/apple.ts');
+    const { createAppleInteractor } = await import('./interactor.ts');
     return createAppleInteractor(device, runner);
   },
   // Reproduces the macOS host fast-path + Apple-simulator branch of the
   // inventory if-chain, reusing the SAME predicate (no divergent copy).
   discoverDevices: async (request: DeviceInventoryRequest) => {
     if (shouldUseHostMacFastPath(request)) {
-      const { listMacosDevices } = await import('../../platforms/apple/os/macos/devices.ts');
+      const { listMacosDevices } = await import('./os/macos/devices.ts');
       return await listMacosDevices();
     }
-    const { listAppleDevices } = await import('../../platforms/apple/core/devices.ts');
+    const { listAppleDevices } = await import('./core/devices.ts');
     return await listAppleDevices({
       simulatorSetPath: request.iosSimulatorSetPath,
       udid: request.udid,
     });
   },
 } as const satisfies PlatformPlugin;
-
-const androidPlugin = {
-  id: 'android',
-  platforms: ['android'],
-  capability: {
-    bucket: 'android',
-    supportsByDefault: { [PUBLIC_COMMANDS.audio]: isAudioProbeSupportedDevice },
-  },
-  // Wraps the Android arm of `resolveLogBackend`: every Android device -> 'android'.
-  appLog: { resolveBackend: () => 'android' },
-  // Wraps the Android arm of `supportsPlatformPerfMetrics`: every Android device
-  // reports perf-metrics support.
-  perf: { supportsMetrics: () => true },
-  createInteractor: async (device: DeviceInfo) => {
-    const { createAndroidInteractor } = await import('../interactors/android.ts');
-    return createAndroidInteractor(device);
-  },
-  discoverDevices: async (request: DeviceInventoryRequest) => {
-    const { listAndroidDevices } = await import('../../platforms/android/devices.ts');
-    return await listAndroidDevices({
-      serialAllowlist: request.androidSerialAllowlist
-        ? new Set(request.androidSerialAllowlist)
-        : undefined,
-    });
-  },
-} as const satisfies PlatformPlugin;
-
-const linuxPlugin = {
-  id: 'linux',
-  platforms: ['linux'],
-  capability: { bucket: 'linux' },
-  createInteractor: async () => {
-    const { createLinuxInteractor } = await import('../interactors/linux.ts');
-    return createLinuxInteractor();
-  },
-  discoverDevices: async () => {
-    const { listLinuxDevices } = await import('../../platforms/linux/devices.ts');
-    return await listLinuxDevices();
-  },
-} as const satisfies PlatformPlugin;
-
-const webPlugin = {
-  id: 'web',
-  platforms: ['web'],
-  capability: { bucket: 'web' },
-  createInteractor: async () => {
-    const { createWebInteractor } = await import('../interactors/web.ts');
-    return createWebInteractor();
-  },
-  // Mirrors the `request.platform === 'web'` branch (the single static device).
-  discoverDevices: async () => [WEB_DESKTOP_DEVICE],
-} as const satisfies PlatformPlugin;
-
-/**
- * The builtin plugins, in `PLATFORMS` order so `registeredPlatforms()` derives
- * the canonical tuple's order (asserted by the parity test).
- */
-export const BUILTIN_PLATFORM_PLUGINS = [
-  applePlugin,
-  androidPlugin,
-  linuxPlugin,
-  webPlugin,
-] as const satisfies readonly PlatformPlugin[];
-
-// The leaf platforms covered by at least one builtin plugin, recovered from the
-// preserved literal `platforms` tuples.
-type CoveredPlatform = (typeof BUILTIN_PLATFORM_PLUGINS)[number]['platforms'][number];
-
-/**
- * Compile-time EXHAUSTIVENESS: a new `Platform` literal added to `PLATFORMS`
- * without a plugin makes `Platform` no longer extend `CoveredPlatform`, so this
- * alias resolves to `false`, violating the `extends true` constraint and failing
- * the build. This is the registry counterpart of the deleted `getInteractor`
- * switch's exhaustive `never` default. (Equivalent in spirit to the §5.1
- * `Object.fromEntries(registeredPlatforms()...) satisfies Record<Platform, true>`
- * sketch, but type-level so it cannot be satisfied vacuously by a runtime map.)
- */
-type AssertTrue<T extends true> = T;
-export type BuiltinPluginsCoverAllPlatforms = AssertTrue<
-  [Platform] extends [CoveredPlatform] ? true : false
->;
-
-let registered = false;
-
-/**
- * Registers every builtin plugin into the shared registry exactly once
- * (idempotent). Called at the top of `core/interactors.ts` so the registry is
- * populated before any `getPlugin` lookup; safe to call again from tests.
- */
-export function registerBuiltinPlatformPlugins(): void {
-  if (registered) return;
-  for (const plugin of BUILTIN_PLATFORM_PLUGINS) {
-    registerPlatformPlugin(plugin);
-  }
-  registered = true;
-}
