+++ src/sdk/index.ts
@@ -2,134 +2,3 @@ export { createAgentDeviceClient } from '../client/client.ts';
 export { createLocalArtifactAdapter } from '../io.ts';
 export { AppError, isAgentDeviceError, normalizeAgentDeviceError } from '../kernel/errors.ts';
 export { centerOfRect } from '../kernel/snapshot.ts';
-
-export type {
-  ArtifactAdapter,
-  ArtifactDescriptor,
-  CreateTempFileOptions,
-  FileInputRef,
-  FileOutputRef,
-  LocalArtifactAdapterOptions,
-  OutputVisibility,
-  ReserveOutputOptions,
-  ReservedOutputFile,
-  ResolveInputOptions,
-  ResolvedInputFile,
-  TemporaryFile,
-} from './io.ts';
-
-export type { AppErrorCode, NormalizedError } from '../kernel/errors.ts';
-
-export type {
-  ReplayTestReporter,
-  ReplayTestReporterContext,
-  ReplayTestReporterFactory,
-  ReplayTestReporterLoadContext,
-} from '../replay/test/reporters/types.ts';
-
-export type { CommandResult } from '../core/command-descriptor/command-result.ts';
-export type { ResponseLevel } from '../kernel/contracts.ts';
-export type { BootCommandResult, ShutdownCommandResult } from '../contracts/device.ts';
-export type { ViewportCommandResult } from '../contracts/viewport.ts';
-export type {
-  CloudArtifact,
-  CloudArtifactAvailability,
-  CloudArtifactKind,
-  CloudArtifactsResult,
-  CloudArtifactsStatus,
-} from '../cloud-artifacts.ts';
-
-export type {
-  AgentDeviceClient,
-  AgentDeviceClientConfig,
-  AgentDeviceCommandClient,
-  AgentDeviceDaemonTransport,
-  AgentDeviceDevice,
-  AgentDeviceIdentifiers,
-  AgentDeviceRequestOverrides,
-  AgentDeviceSelectionOptions,
-  AgentDeviceSession,
-  AgentDeviceSessionDevice,
-  AlertAction,
-  AlertCommandOptions,
-  AlertInfo,
-  AlertPlatform,
-  AlertSource,
-  AppCloseOptions,
-  AppCloseResult,
-  AppDeployOptions,
-  AppDeployResult,
-  AppInstallFromSourceOptions,
-  AppInstallFromSourceResult,
-  AppListOptions,
-  AppOpenOptions,
-  AppOpenResult,
-  AppPushOptions,
-  AppStateCommandOptions,
-  AppStateCommandResult,
-  AppSwitcherCommandOptions,
-  AppSwitcherCommandResult,
-  AppTriggerEventOptions,
-  BackCommandOptions,
-  BackCommandResult,
-  BatchRunOptions,
-  BatchRunResult,
-  BatchStep,
-  CaptureDiffOptions,
-  CaptureScreenshotOptions,
-  CaptureScreenshotResult,
-  CaptureSnapshotOptions,
-  CaptureSnapshotResult,
-  ClickOptions,
-  ClipboardCommandOptions,
-  ClipboardCommandResult,
-  CloudArtifactsOptions,
-  CommandRequestResult,
-  DeviceBootOptions,
-  DeviceShutdownOptions,
-  ElementTarget,
-  FillOptions,
-  FindLocator,
-  FindOptions,
-  FocusOptions,
-  GetOptions,
-  HomeCommandResult,
-  InteractionTarget,
-  IsOptions,
-  KeyboardCommandOptions,
-  KeyboardCommandResult,
-  LogsOptions,
-  LongPressOptions,
-  MaterializationReleaseOptions,
-  MaterializationReleaseResult,
-  MetroPrepareOptions,
-  MetroPrepareResult,
-  NetworkOptions,
-  PerfOptions,
-  PermissionTarget,
-  PinchOptions,
-  PressOptions,
-  RecordOptions,
-  ReplayRunOptions,
-  ReplayTestOptions,
-  RotateCommandOptions,
-  RotateCommandResult,
-  ScrollOptions,
-  SessionCloseResult,
-  SettingsUpdateOptions,
-  StartupPerfSample,
-  SwipeOptions,
-  TargetShutdownResult,
-  TraceOptions,
-  TypeTextOptions,
-  WaitCommandOptions,
-} from '../client/client.ts';
-
-export type {
-  Point,
-  Rect,
-  ScreenshotOverlayRef,
-  SnapshotNode,
-  SnapshotVisibility,
-  SnapshotVisibilityReason,
-} from '../kernel/snapshot.ts';
